+++ kamu/__init__.py
@@ -1,6 +1,6 @@
 from ._connection import KamuConnection
 
-__version__ = "0.1.2"
+__version__ = "0.2.0"
 
 
 def connect(url, engine=None):